split instrumentation
